added query command module '?' can be used to store the current turtle state matrix in bpy.context.scene.turtleStateQueryResults from where it can be accessed from within L-Py code via import bpy.
by comparing the turtle state to other variables from bpy.context.scene
this allows the modelling of environmental interaction of plants
thanks to integration of L-Py with Blender via Lindenmaker.

details on usage are documented in README file.

to implement this feature, some significant changes had to be made:
- there is now a Turtle base class that does no drawing
but can move and do queries, as well as a DrawingTurtle subclass.
- to query the turtle state after each L-Py production step,
we can no longer do all production steps at once in L-Py,
but need to do them stepwise with LSystem.derivationlength=1
(we can still do that multiple times however),
since we need to run the turtle interpretation after each step
to query the turtle state we need in the next L-Py production step.
to avoid the overhead of drawing, we use the simple
Turtle base class for this purpose with negligible overhead.